// src/app/model/auto.ts
export interface auto{
    id?: string
    alias?: string;
    name?: string;
    image_url?: string;
    is_closed?: boolean;
    url?: string;
    review_count?: number,
    categories?: Array<Object>;
    rating?: number;
    coordinates?: Object;
    display_phone?: string;
    distance?: number;
    location?: Object;
    price?: string;
    transactions?: Array<string>;
    display_address?: Array<string>;
    terms?: Array<Object>;
    businesses?: Array<Object>;
    title?: string;
    

  }

  